Remove memset pattern when initializing transaction set
Problem: The call to memset is unnecessary, as the following call to
init_all_transactions() properly zeros out the transaction set block.
Its also disruptive, as we are accidentally zeroing out the memory
capturing our timer object such that we're no longer able to cancel
pending timers when we re-intialize transactions. This is making it so
disconnection events can't close out pending timers. When a
disconnection is closely followed by a reconnection, as in a connection
collision, the un-cancelled timer can pop and impact ongoing
transactions. At worst, this is a valid event registration, which causes
us to stop registering for the given event, as we assume the request
timed out and isn't supported.

Solution: Remove the call to memset and rely on init_all_transactions to
properly cancel timers and set memory. Add logging to better track our
transaction labels.

Bug: 344045293
Flag: EXEMPT, no logical change, removal of repeat code
Test: m com.android.btservices
Change-Id: I60250e66783d62efb231a77b3970141a0cac31d9