btif_av: Make public the methods from btav_source_interface_t
The interfaces btif_av_get_src_interface() and
get_profile_interface(("a2dp") are not used
except in the JNI layer.

The call to get_profile_interface in com_android_bluetooth_a2dp
is removed in order to de-obfuscate the call graph for a2dp.

The inferface type and declaration is also removed now that it is
unused.

Bug: 331817295
Test: m com.android.btservices
Flag: EXEMPT, no logical change
Change-Id: I358b84fb3a3d479965185f8537c240a28c6c06f6