Remove auto_connect_on_hfp_when_no_a2dp_device flag
This flag was added in Android U to enable auto-connecting to HFP devices when there is no A2DP device to connect to. This flag has been enabled by default since 24Q2, and there have been no reports of any issues with it. Therefore, it is safe to remove this flag.

Change-Id: Ic73e00116cc1d766d997eb76ba89850bbc2f6569
Test: atest PhonePolicyTest
Bug: 347316578
Flag: exempt, remove nextfood flag