LeAudio: Fix broadcast cleanup issue
During BT stack shutdown the unicast profile calls Stop() on the
CodecManager singleton, and after that brodcast profile tries to call
`UpdateActiveBroadcastAudioHalClient()` which fails. Since the run time
of codec manager is controlled by the unicast profile, we should not
assert in the broadcaster profile shutdown if the call fails, since the
broadcaster profile is shutting down anyway, and this function call
failure is harmless.

Bug: 347845613
Flag: EXEMPT; Trivial fix covered with unit test
Test: atest bluetooth_test_broadcaster && mmm packages/modules/Bluetooth && start LE audio broadcast, then turn BT off
Change-Id: Ic3bd86069354bde6189c1ab892f51d3f6bc37768