Return null instead of empty list when AVRCP CT service isn't ready
Media Center and other browse clients are expected to be able to handle
null and empty folder item responses, and both are indicative of
"there's no results here." However, null results is indicative of an
error case which we can help them recover from (connect a device, but
turn on BT first). Our playback state already indicates an error state
when BT is off, so this makes our browse tree match with that. This also
lets the AAOS Media Center app on cars surface a full screen error
resolution activity that can be properly dismissed when a device
connects.

Bug: 348502174
Flag: EXEMPT, no logic change, only a default return value change
Test: atest AvrcpControllerStateMachineTest.java -- tests the browser
Test: service interface in addition to the state machine
Change-Id: I0eff67d89e7c194fb3f5ac5e52a4d98ae9ed9732