SystemServer: SatelliteMode bypass thread jump
The callback from the satellite listener and the airplane mode are
already executed in the correct context.
It is not necessary to send a message to disable, instead call disable
directly.

Note that this was already the case to finalize the disable of the BLE
mode

Bug: 349428963
Test: Manual:
      Trigger satellite mode while bluetooth is on and observe logs
Flag: Exempt: very specific usecase that is entirely tested
Change-Id: I298f71e0c24c8595692290dea9379c1c22fcc44d