InCallService: Disconnecting as disconnected
During hang up of a call, Telecom update the call state from
Call.STATE_ACTIVE to Call.STATE_DISCONNECTING and propagate it to
BluetoothInCallService.

And BluetoothInCallService update the phone state with numActive=0 and
callState=idle to HFP, this result in SCO link disconnected before the
call end.

The solution is to treat Call.STATE_DISCONNECTING as disconnected, so
HFP does not disconnect the SCO link when call disconnected.

Bug: 328847580
Fix: 328847580
Test: None
Flag: Exempt convergence code
Change-Id: I06e05914de5911ddf9ce9d7a6e383d512375ef26