floss: Guarantee the execution order of GATT client callbacks
Originally, the callbacks are sent in an async thread, which means it
could be interrupted by the other tasks, and thus their order after
sent is not guaranteed to be the same as before.
This patch pushes the callbacks into a queue and makes sure the pop+send
operation is not interrupted, this shall guarantee the order after sent.

Bug: 348270691
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: manual test with an LE mouse and BAS works
Flag: EXEMPT, Floss-only changes
Change-Id: I5eeec16197b4d3b41e4feba9e76cd44181adefb7