system/bta: Use correct avdt handle for bta_av_setconfig_rej
bta_av_adjust_seps_idx fails when the remote configuration
is invalid and p_scb->avdt_handle is not configured as it should be.
This results in p_scb->avdt_handle being invalid and the
AVDTP Set Configuration Reject response not being sent.

This change removes the call to bta_av_adjust_seps_idx
which is only useful when accepting the configuration, and
directly uses the input handle.

Bug: 336232163
Bug: 338139069
Test: m com.android.btservices
Test: atest pts-bot:A2DP/SNK/AVP
Flag: com.android.bluetooth.flags.avdtp_error_codes
Change-Id: I79eed6b1d453b2e3c1ee6dc40d5925d2a2b7a837