[le audio] Avoid clearing cached sinks with multiple stream requests
Refactoring the usage of mPausedBroadcastSink, and add some fixes.
If the device is in handover state due to previous local stream request,
we don't need to clear the cached in following requests.

This is also to avoid the race condition that some sinks are removing
the source and we've cleared the previous cached sinks.

Bug: 355479593
Test: atest BassClientSericeTest
Test: manual test private broadcast with multiple unicast stream
requests
Flag: Exempt; trivial change covered with unit and manual test

Change-Id: Ie1ca9689f7155684de511a7c2692ef528e2e47d9