Notification: Allow manual override of the count
When the count is manually reset to 0, for tests purposes, the date can
still be set and this can lead to the notification not being displayed.

It is simpler to modify the test script step to include it, but I
believe that the issue maybe be present for multiple test suite of
multiple partner. Also, we need to think about an eventual power user
that would like to use that.

The date check is moving to be done only if the count indicate that
there is already at lest one notification that was sent.

Note that factory reset of Bluetooth should in the long run restore all
this parameter.

Bug: 355602225
Fix: 355602225
Test: None
Flag: Exempt trivial change
Change-Id: Ib5c07a06195998b6df4f077fccde18e6256e3d78