HidHost: Remove quiet mode dead code
mTarget is null and never set in the code. Exiting the quiet mode does
nothing.
Removing this whole code since the potential call to offToBleOn is very
scary (even if impossible to trigger, since when bluetooth is off, there
is no HidHost code that run)

Bug: 311772251
Test: None
Flag: Exempt dead code removal
Change-Id: I204cd8c639b9a21f44ca3b53382f0f1808e2ef51