Floss: Increase the stack start timeout to 12 seconds
In Linux kernel, all Bluetooth drivers do HCI Reset when the Bluetooth
socket is closed. This help to save power and keep the host and
controller states synchronized. However, it is also observed that the
HCI Reset could sometimes take a long time, which will in turn cause the
next stack initialization to timeout (3 seconds) because the driver is
still waiting for the HCI Reset to complete.

To mitigate the issue, this CL sets the start timeout to be 12 seconds,
which is longer than the driver timeout to wait for the HCI Reset.

Bug: 357792559
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: Manually modify the kernel so that the shutdown process is always
      timeout, then do usbreset. Verify that there is no crash.
Flag: EXEMPT, Floss-only changes

Change-Id: I54a1dadbe593dcf295fe3ae1d146ad98d136d863