Skip AVRCP PTS-Bot tests that require a flag
These two tests need to run with the browsing refactor flag as the MMI
description does not fit anymore. There is a bug with presubmit not
being able to enable the flag properly right now so we have to skip
them until this is fixed and re-enable them in the flag section.

Bug: 359175417
Flag: Exempt testing
Test: atest pts-bot:AVRCP
Change-Id: Ic953998dba5c45dfcc3cfb59adea2ab02eaf0928